# openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Task API",
        "version": "1.0.0",
        "description": "A simple CRUD API built with Express"
    },
    "paths": {
        "/": {
            "get": {
                "summary": "Get API information",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },

        "/health": {
            "get": {
                "summary": "Health Check",
                "responses": {
                    "200": {
                        "description": "Server is running"
                    }
                }
            }
        },

        "/tasks": {
            "get": {
                "summary": "Get all tasks",
                "responses": {
                    "200": {
                        "description": "List of tasks"
                    }
                }
            },

            "post": {
                "summary": "Create a new task",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Task created"
                    },
                    "400": {
                        "description": "Invalid request"
                    }
                }
            }
        },

        "/tasks/{id}": {
            "get": {
                "summary": "Get task by ID",
                "parameters": [{
                    "name": "id",
                    "in": "path",
                    "required": true,
                    "schema": {
                        "type": "integer"
                    }
                }],
                "responses": {
                    "200": {
                        "description": "Task found"
                    },
                    "404": {
                        "description": "Task not found"
                    }
                }
            },

            "put": {
                "summary": "Update a task",
                "parameters": [{
                    "name": "id",
                    "in": "path",
                    "required": true,
                    "schema": {
                        "type": "integer"
                    }
                }],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "done": {
                                        "type": "boolean"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Task updated"
                    },
                    "400": {
                        "description": "Invalid request"
                    },
                    "404": {
                        "description": "Task not found"
                    }
                }
            },

            "delete": {
                "summary": "Delete a task",
                "parameters": [{
                    "name": "id",
                    "in": "path",
                    "required": true,
                    "schema": {
                        "type": "integer"
                    }
                }],
                "responses": {
                    "204": {
                        "description": "Task deleted"
                    },
                    "404": {
                        "description": "Task not found"
                    }
                }
            }
        }
    }
}